controller: Added controller for auth route
Separating auth route into its own controller will help modularity and easy
testing of business logic.
